+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Fixed
+- chroot: fix tedious shell glitches in clara chroot using real shell
 ## Added
 - ipmi: provide an ssh_jump_host configuration option for connect